Add files via upload
I updated the button style, to edit it, you will need to install the krypton component system into VS, but if not, the files are already created for you to use the generator. If you decide to edit the program and build the solution it will pop out the exe and a dll file. If you want it to only have the exe file you will need an external program called ILMerge which will allow you to bind the exe and dll files together. the krypton pack is here https://github.com/ComponentFactory/Krypton and the ILMerge is here https://wvd-vegt.bitbucket.io/ilmergegui.html